feat: re-introduce docs package
The `mermaid` plugin for `mkdocs` likes to source `mermaidjs` remotely at run/build time. You can resolve it locally instead, which is what I've configured.

I added a `mermaid` package to this flake which captures `mermaidjs`, and have configured the `devshell` and the `docs` package to symlink it into `./docs/js/mermaid`.

Signed-off-by: Brian McGee <brian@bmcgee.ie>